Added first implementation of First Person Camera
Updated the project and the code to use the supplied camera class in the
library.  Seems to work alright for now.  Hold left mouse button and
move mouse to change view; use wasd to move the position of the camera.
Amount of movement is correlated to the distance between camera and
surface of earth.  (as it exists now).  closes #1