Working towards modular project
Getting the dll-based separate projects to work with DX11 & VS 2013
doesn't seem to be possible for now.  As such, I've decided to use a
single project with the relevant subsystems broken up into separate
folders & namespaces.  This allows us to work on the code in a modular
manner and will hopefully make transformation easier once we figure out
how to do the dll-based approach.